requirements.txt
streamlit
google-generativeai
gtts
pandas